+++ src/main/java/org/markdownwriterfx/options/OptionsDialog.java
@@ -28,7 +28,11 @@
 package org.markdownwriterfx.options;
 
 import javafx.event.ActionEvent;
+import javafx.scene.Node;
+import javafx.scene.control.ButtonBar.ButtonData;
+import javafx.scene.control.ButtonBar;
 import javafx.scene.control.ButtonType;
+import javafx.scene.control.CheckBox;
 import javafx.scene.control.Dialog;
 import javafx.scene.control.DialogPane;
 import javafx.scene.control.Tab;
@@ -36,6 +40,7 @@ import javafx.scene.control.TabPane;
 import javafx.stage.Window;
 import org.markdownwriterfx.MarkdownWriterFXApp;
 import org.markdownwriterfx.Messages;
+import org.markdownwriterfx.projects.ProjectManager;
 import org.markdownwriterfx.util.Utils;
 
 /**
@@ -55,12 +60,33 @@ public class OptionsDialog
 
 		tabPane.getStyleClass().add(TabPane.STYLE_CLASS_FLOATING);
 
+		// add "Store in project" checkbox to buttonbar
+		boolean oldStoreInProject = Options.isStoreInProject();
+		ButtonType storeInProjectButtonType = new ButtonType(Messages.get("OptionsDialog.storeInProject.text"), ButtonData.LEFT);
+		setDialogPane(new DialogPane() {
+			@Override
+			protected Node createButton(ButtonType buttonType) {
+				if (buttonType == storeInProjectButtonType) {
+					CheckBox storeInProjectButton = new CheckBox(buttonType.getText());
+					ButtonBar.setButtonData(storeInProjectButton, buttonType.getButtonData());
+					storeInProjectButton.setSelected(oldStoreInProject);
+					storeInProjectButton.setDisable(ProjectManager.getActiveProject() == null);
+					return storeInProjectButton;
+				}
+				return super.createButton(buttonType);
+			}
+		});
+
 		DialogPane dialogPane = getDialogPane();
 		dialogPane.setContent(tabPane);
-		dialogPane.getButtonTypes().addAll(ButtonType.OK, ButtonType.CANCEL);
+		dialogPane.getButtonTypes().addAll(storeInProjectButtonType, ButtonType.OK, ButtonType.CANCEL);
 
 		// save options on OK clicked
 		dialogPane.lookupButton(ButtonType.OK).addEventHandler(ActionEvent.ACTION, e -> {
+			boolean newStoreInProject = ((CheckBox)dialogPane.lookupButton(storeInProjectButtonType)).isSelected();
+			if (newStoreInProject != oldStoreInProject)
+				Options.storeInProject(newStoreInProject);
+
 			save();
 			e.consume();
 		});